==45== Uppload up status when weather station turned on, some due compatibility, turn off some code optimizations. Seems not to hang any more when I2C freezes.

+++ WeatherStationCommon/sensors/SensorFlags.cpp
@@ -18,7 +18,7 @@ BitBool<WeatherStation::SensorValueId::SensorsEnumSize> WeatherStation::SensorFl
 void WeatherStation::SensorFlags::writeFlag(SensorValueId sensorId, bool value,
                                         ReadingUploader uploader) {
     if (uploader >= ReadingUploader::ReadingUploaderSize){
-        for (int i = 0; i < ReadingUploader::ReadingUploaderSize; i++) {
+        for (uint8_t i = 0; i < ReadingUploader::ReadingUploaderSize; i++) {
             sensorFlags[i][sensorId] = value;
         }
     } else {
